pyproject.toml
[tool.poetry]
name = "telegram-dynamic-nicknames"
version = "1.0.0"
description = "Script to generate dynamic Telegram nicknames"
authors = ["Slava Wedyarit <wedyarit@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Wedyarit/telegram-dynamic-nicknames"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Typing :: Typed",
    "Private :: Do Not Upload",
]

[tool.poetry.dependencies]
python = "^3.10"
telethon = "^1.25.4"
python-dotenv = "^0.21.0"